hotfix broken import stable_baselines (#439)
* test_no_mpi: Use rreload to catch ImportError

`reload` and `import` are not sufficient.

* Move ddpg.noise to common.noise

* rreload fix

* Replace ddpg.noise references with common.noise

* Update changelog

* Revert "Update changelog"

This reverts commit 2316ea224e6a7ddb5b183259ef432f4190d00290.

* Revert "Replace ddpg.noise references with common.noise"

This reverts commit 2a6987fb2aa0d5b6a461751f80128a4fd542fe81.

* Simpler method to re-import stable baselines

* Linting

* Add sb.ddpg.noise.py stub

* Try disable pylint unused import

* fix typo

* Another typo

* trigger ci

(Travis CI was down)

* Restore old sys.modules to try and fix isinstance error

* Linting

* update changelog
